+++ src/sql/exceptions.py
@@ -38,6 +38,7 @@ MissingPackageError = exception_factory("MissingPackageError")
 TypeError = exception_factory("TypeError")
 RuntimeError = exception_factory("RuntimeError")
 ValueError = exception_factory("ValueError")
+KeyError = exception_factory("KeyError")
 FileNotFoundError = exception_factory("FileNotFoundError")
 NotImplementedError = exception_factory("NotImplementedError")