+++ src/tests/test_magic.py
@@ -1,3 +1,4 @@
+from unittest.mock import ANY
 import uuid
 import logging
 import platform
@@ -20,6 +21,7 @@ from sql.magic import SqlMagic
 from sql.run.resultset import ResultSet
 from sql import magic
 from sql.warnings import JupySQLQuotedNamedParametersWarning
+
 
 from conftest import runsql
 from sql.connection import PLOOMBER_DOCS_LINK_STR
@@ -1770,3 +1772,180 @@ def test_persist_uses_error_handling_method(ip, monkeypatch, cell):
     # ensure this got called because this function handles several sqlalchemy edge
     # cases
     execute_with_error_handling_mock.assert_called_once()
+
+
+def test_error_when_using_section_argument_but_dsn_is_missing(ip_empty, tmp_empty):
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'path/to/connections.ini'")
+
+    with pytest.raises(UsageError) as excinfo:
+        ip_empty.run_cell("%sql --section some_section")
+
+    assert excinfo.value.error_type == "FileNotFoundError"
+    assert "%config SqlMagic.dsn_filename ('path/to/connections.ini') not found" in str(
+        excinfo.value
+    )
+
+
+def test_error_when_using_section_argument_but_dsn_section_is_missing(
+    ip_empty, tmp_empty
+):
+    Path("connections.ini").write_text(
+        """
+[section]
+key = value
+"""
+    )
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    with pytest.raises(UsageError) as excinfo:
+        ip_empty.run_cell("%sql --section another_section")
+
+    assert excinfo.value.error_type == "KeyError"
+
+    message = (
+        "The section 'another_section' does not exist in the "
+        "connections file 'connections.ini'"
+    )
+    assert message in str(excinfo.value)
+
+
+def test_error_when_using_section_argument_but_keys_are_invalid(ip_empty, tmp_empty):
+    Path("connections.ini").write_text(
+        """
+[section]
+key = value
+"""
+    )
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    with pytest.raises(UsageError) as excinfo:
+        ip_empty.run_cell("%sql --section section")
+
+    assert excinfo.value.error_type == "TypeError"
+
+    message = "%config SqlMagic.dsn_filename ('connections.ini') is invalid"
+    assert message in str(excinfo.value)
+
+
+def test_error_when_using_section_argument_but_values_are_invalid(ip_empty, tmp_empty):
+    Path("connections.ini").write_text(
+        """
+[section]
+drivername = not-a-driver
+"""
+    )
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    with pytest.raises(UsageError) as excinfo:
+        ip_empty.run_cell("%sql --section section")
+
+    message = "Could not parse SQLAlchemy URL from string 'not-a-driver://'"
+    assert message in str(excinfo.value)
+
+
+def test_error_when_using_section_argument_and_alias(ip_empty, tmp_empty):
+    Path("connections.ini").write_text(
+        """
+[duck]
+drivername = duckdb
+"""
+    )
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    with pytest.raises(UsageError) as excinfo:
+        ip_empty.run_cell("%sql --section duck --alias stuff")
+
+    assert excinfo.value.error_type == "UsageError"
+
+    message = "Cannot use --section with --alias"
+    assert message in str(excinfo.value)
+
+
+def test_connect_to_db_in_connections_file_using_section_argument(ip_empty, tmp_empty):
+    Path("connections.ini").write_text(
+        """
+[duck]
+drivername = duckdb
+"""
+    )
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    ip_empty.run_cell("%sql --section duck")
+
+    conns = ConnectionManager.connections
+    assert conns == {"duck": ANY}
+
+
+def test_connect_to_db_in_connections_file_using_section_name_between_square_brackets(
+    ip_empty, tmp_empty
+):
+    Path("connections.ini").write_text(
+        """
+[duck]
+drivername = duckdb
+"""
+    )
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    with pytest.warns(FutureWarning) as record:
+        ip_empty.run_cell("%sql [duck]")
+
+    assert "Starting connections with: %sql [section_name] is deprecated" in str(
+        record[0].message
+    )
+    assert len(record) == 1
+    conns = ConnectionManager.connections
+    assert conns == {"duckdb://": ANY}
+
+
+@pytest.mark.parametrize(
+    "content, error_type, error_detail",
+    [
+        (
+            """
+[duck]
+drivername = duckdb
+
+[duck]
+drivername = duckdb
+""",
+            "DuplicateSectionError",
+            "section 'duck' already exists",
+        ),
+        (
+            """
+[duck]
+drivername = duckdb
+drivername = duckdb
+""",
+            "DuplicateOptionError",
+            "option 'drivername' in section 'duck' already exists",
+        ),
+    ],
+    ids=[
+        "duplicate-section",
+        "duplicate-key",
+    ],
+)
+def test_error_when_ini_file_is_corrupted(
+    ip_empty, tmp_empty, content, error_type, error_detail
+):
+    Path("connections.ini").write_text(content)
+
+    ip_empty.run_cell("%config SqlMagic.dsn_filename = 'connections.ini'")
+
+    with pytest.raises(UsageError) as excinfo:
+        ip_empty.run_cell("%sql --section duck")
+
+    assert "An error happened when loading your %config SqlMagic.dsn_filename" in str(
+        excinfo.value
+    )
+
+    assert error_type in str(excinfo.value)
+    assert error_detail in str(excinfo.value)